Test script update - numbering, postconditions
