+++ test_scripts/Polices/build_options/011_ATF_P_TC_PTS_Define_URL_to_send_PTS_HTTP.lua
@@ -1,0 +1,75 @@
+---------------------------------------------------------------------------------------------
+-- Requirements summary:
+-- [PolicyTableUpdate] Define the URL(s) the PTS will be sent to
+--
+-- Description:
+-- To get the urls PTS should be transfered to, Policies manager must refer PTS "endpoints" section,
+-- key "0x07" for the appropriate <app id> which was chosen for PTS transferring
+-- 1. Used preconditions
+-- SDL is built with "-DEXTENDED_POLICY: HTTP" flag
+-- 2. Performed steps
+-- Application is registered.
+-- PTU is requested.
+-- SDL->HMI: SDL.OnStatusUpdate(UPDATE_NEEDED)
+-- SDL->HMI:SDL.PolicyUpdate(file, timeout, retry[])
+-- HMI -> SDL: SDL.GetURLs (<service>)
+--
+-- Expected result:
+-- SDL.GetURLs({urls[] = default}, (<urls>, appID))
+-- SDL-> <app ID> ->OnSystemRequest(params, url, timeout)
+---------------------------------------------------------------------------------------------
+
+--[[ General configuration parameters ]]
+config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
+
+--[[ Required Shared libraries ]]
+local commonSteps = require('user_modules/shared_testcases/commonSteps')
+local commonFunctions = require('user_modules/shared_testcases/commonFunctions')
+local testCasesForPolicyTableSnapshot = require('user_modules/shared_testcases/testCasesForPolicyTableSnapshot')
+
+--[[ General Precondition before ATF start ]]
+commonSteps:DeleteLogsFileAndPolicyTable()
+--TODO: Should be removed when issue: "ATF does not stop HB timers by closing session and connection" is fixed
+config.defaultProtocolVersion = 2
+
+--[[ General Settings for configuration ]]
+Test = require('connecttest')
+require('cardinalities')
+require('user_modules/AppTypes')
+
+--[[ Test ]]
+commonFunctions:newTestCasesGroup("Test")
+
+function Test:TestStep_PolicyManager_sends_PTS_to_HMI()
+  local endpoints = {}
+  local is_app_esxist = false
+
+  for i = 1, #testCasesForPolicyTableSnapshot.pts_endpoints do
+    -- Take default endpoints
+    if (testCasesForPolicyTableSnapshot.pts_endpoints[i].service == "0x07") then
+      endpoints[#endpoints + 1] = { url = testCasesForPolicyTableSnapshot.pts_endpoints[i].value, appID = nil}
+    end
+    -- Take appID endpoints
+    if (testCasesForPolicyTableSnapshot.pts_endpoints[i].service == "app1") then
+      endpoints[#endpoints + 1] = { url = testCasesForPolicyTableSnapshot.pts_endpoints[i].value, appID = testCasesForPolicyTableSnapshot.pts_endpoints[i].appID}
+      is_app_esxist = true
+    end
+  end
+
+  local RequestId_GetUrls = self.hmiConnection:SendRequest("SDL.GetURLS", { service = 7 })
+
+  EXPECT_HMIRESPONSE(RequestId_GetUrls,{result = {code = 0, method = "SDL.GetURLS", urls = endpoints} } )
+  :Do(function(_,_)
+      if(is_app_esxist == false) then
+        self:FailTestCase("endpoints for application doesn't exist!")
+      end
+    end)
+end
+
+--[[ Postconditions ]]
+
+function Test.Postcondition_Stop_SDL()
+  StopSDL()
+end
+
+return Test
