--- test_scripts/Polices/build_options/015_ATF_P_TC_PTS_Define_URL_to_send_PTS_HTTP.lua
@@ -1,131 +1,0 @@
----------------------------------------------------------------------------------------------
--- Requirements summary:
--- [PolicyTableUpdate] Define the URL(s) the PTS will be sent to
---
--- Description:
--- SDL should request PTU in case user requests PTU
--- 1. Used preconditions
--- SDL is built with "-DEXTENDED_POLICY: HTTP" flag
--- 2. Performed steps
--- Application is registered.
--- PTU is requested.
--- SDL->HMI: SDL.OnStatusUpdate(UPDATE_NEEDED)
--- SDL->HMI:SDL.PolicyUpdate(file, timeout, retry[])
--- HMI -> SDL: SDL.GetURLs (<service>)
---
--- Expected result:
--- SDL->HMI: SDL.OnStatusUpdate(UPDATE_NEEDED)
--- SDL->HMI:SDL.PolicyUpdate(file, timeout, retry[])
----------------------------------------------------------------------------------------------
-
---[[ General configuration parameters ]]
-config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
-
---[[ Required Shared libraries ]]
-local commonSteps = require('user_modules/shared_testcases/commonSteps')
-local commonFunctions = require('user_modules/shared_testcases/commonFunctions')
-local commonPreconditions = require('user_modules/shared_testcases/commonPreconditions')
-local testCasesForPolicyTableSnapshot = require('user_modules/shared_testcases/testCasesForPolicyTableSnapshot')
-
---[[ General Precondition before ATF start ]]
--- commonFunctions:SDLForceStop()
-commonSteps:DeleteLogsFileAndPolicyTable()
-commonPreconditions:Connecttest_without_ExitBySDLDisconnect_WithoutOpenConnectionRegisterApp("connecttest_RAI.lua")
-
---ToDo: should be removed when issue: "ATF does not stop HB timers by closing session and connection" is fixed
-config.defaultProtocolVersion = 2
-
---[[ General Settings for configuration ]]
-Test = require('user_modules/connecttest_RAI')
-require('cardinalities')
-require('user_modules/AppTypes')
-local mobile_session = require('mobile_session')
-
---[[ Preconditions ]]
-commonFunctions:newTestCasesGroup("Preconditions")
-function Test.Precondition_remove_user_connecttest()
-  os.execute( "rm -f ./user_modules/connecttest_RAI.lua" )
-end
-
---[[ Test ]]
-commonFunctions:newTestCasesGroup("Test")
-function Test:TestStep_ConnectMobile()
-  self:connectMobile()
-end
-
-function Test:TestStep_StartNewSession()
-  self.mobileSession = mobile_session.MobileSession( self, self.mobileConnection)
-  self.mobileSession:StartService(7)
-end
-
---TODO(mmihaylova-banska): Function still not implemented
-function Test:TestStep_PolicyManager_sends_PTS_to_HMI()
-  local is_test_fail = false
-  local correlationId = self.mobileSession:SendRPC("RegisterAppInterface", config.application1.registerAppInterfaceParams)
-
-  EXPECT_HMINOTIFICATION("BasicCommunication.OnAppRegistered", { application = { appName = config.application1.appName } })
-  :Do(function(_,data)
-      local hmi_app_id = data.params.application.appID
-      EXPECT_HMINOTIFICATION("SDL.OnStatusUpdate", {status = "UPDATE_NEEDED"})
-
-      testCasesForPolicyTableSnapshot:verify_PTS(true,
-        {config.application1.registerAppInterfaceParams.appID},
-        {config.deviceMAC},
-        {hmi_app_id})
-
-      local timeout_after_x_seconds = testCasesForPolicyTableSnapshot:get_data_from_PTS("module_config.timeout_after_x_seconds")
-      local seconds_between_retries = {}
-
-      for i = 1, #testCasesForPolicyTableSnapshot.pts_seconds_between_retries do
-        seconds_between_retries[i] = testCasesForPolicyTableSnapshot.pts_seconds_between_retries[i].value
-      end
-
-      --TODO(mmihaylova-banska): Function still not implemented
-      --local SystemFilesPath = commonSteps:get_data_form_SDL_ini("SystemFilesPath")
-      local SystemFilesPath = "/tmp/fs/mp/images/ivsu_cache/"
-      local file_pts = SystemFilesPath.."sdl_snapshot.json"
-
-      EXPECT_HMICALL("BasicCommunication.PolicyUpdate", { file = file_pts, timeout = timeout_pts, retry = seconds_between_retries})
-      :Do(function(_,data1)
-          self.hmiConnection:SendResponse(data1.id, data1.method, "SUCCESS", {})
-        end)
-
-      local endpoints = {}
-      local is_app_esxist = false
-
-      for i = 1, #testCasesForPolicyTableSnapshot.pts_endpoints do
-        if (testCasesForPolicyTableSnapshot.pts_endpoints[i].service == "0x07") then
-          endpoints[#endpoints + 1] = { url = testCasesForPolicyTableSnapshot.pts_endpoints[i].value, appID = nil}
-        end
-
-        if (testCasesForPolicyTableSnapshot.pts_endpoints[i].service == "app1") then
-          endpoints[#endpoints + 1] = { url = testCasesForPolicyTableSnapshot.pts_endpoints[i].value, appID = testCasesForPolicyTableSnapshot.pts_endpoints[i].appID}
-          is_app_esxist = true
-        end
-      end
-
-      local RequestId_GetUrls = self.hmiConnection:SendRequest("SDL.GetURLS", { service = 7 })
-
-      EXPECT_HMIRESPONSE(RequestId_GetUrls,{result = {code = 0, method = "SDL.GetURLS", urls = endpoints} } )
-      :Do(function(_,data)
-          if(is_app_esxist == false) then
-            commonFunctions:printError("endpoints for application doesn't exist!")
-            is_test_fail = true
-          end
-        end)
-
-    end)
-  self.mobileSession:ExpectResponse(correlationId, { success = true, resultCode = "SUCCESS"})
-
-  if(is_test_fail == true) then
-    self:FailTestCase("Test is FAILED. See prints.")
-  end
-end
-
---[[ Postconditions ]]
-commonFunctions:newTestCasesGroup("Postconditions")
-function Test:Postcondition_Force_Stop_SDL()
-  commonFunctions:SDLForceStop(self)
-end
-
-return Test
